\name{DH_magnify}
\alias{DH_magnify}
\title{Magnify a plot}
\description{Zoom by user control in a plot graph already opened}
\usage{DH_magnify()}
\details{
This option allows to zoom a plot. When clicking on the menu item, the list of instruction
 (both mouse and keyboard can be used) appears on the R console:\cr
\cr
\bold{Mouse:}\cr
Right to zoom in, Middle or Hold Left + click right to zoom out
Hold left mouse button to move\cr
\cr
\bold{Keyboard:}\cr
\tabular{rl}{
Left/Right (h/l) \tab move left/right \cr
Up/Down (k/j) \tab move up/down \cr
[+ or i / - or o \tab zoom in/out \cr
L/H  \tab zoom in/out (x-axis only) \cr
K/J \tab zoom in/out (y-axis only) \cr
p \tab print to file \cr
r \tab reset limits \cr
s \tab show limits \cr
q \tab on the graphic window to quit \cr
}
}
\value{None: plot should change}
\references{}
\author{Riccardo Leardi and Gianmarco Polotti with contributions from Giorgio Marubini.Gruppo di Chemiometria (Divisione di Chimica Analitica della Societa' Chimica Italiana)}
\note{}
\seealso{}
\examples{}
\keyword{ magnify}
